requirements.txt
openai-whisper
sounddevice
scipy
numpy
keyboard
